+++ include/Enums.hpp
@@ -1,0 +1,30 @@
+#pragma once
+
+/// The types of icons and colors that can be randomized.
+enum class RandomizeType {
+    Color1 = 0,
+    Color2 = 1,
+    GlowColor = 2,
+    Glow = 3,
+    Cube = 4,
+    Ship = 5,
+    Ball = 6,
+    Ufo = 7,
+    Wave = 8,
+    Robot = 9,
+    Spider = 10,
+    Swing = 11,
+    Jetpack = 12,
+    DeathEffect = 13,
+    Trail = 14,
+    ShipFire = 15,
+    Animation = 16,
+    Explode = 17
+};
+
+/// The types of groups of icons and colors that can be randomized.
+enum class RandomizeAllType {
+    Icons = 0,
+    Special = 1,
+    Colors = 2
+};