+++ src/IconRandomizer.cpp
@@ -1,10 +1,12 @@
 #include "Internal.hpp"
 #include <Geode/binding/GameManager.hpp>
 #include <Geode/binding/GameStatsManager.hpp>
+#include <Geode/loader/GameEvent.hpp>
+#include <Geode/loader/Mod.hpp>
+#include <Geode/utils/random.hpp>
 #define MORE_ICONS_EVENTS
-#include <hiimjustin000.more_icons/include/MoreIconsV2.hpp>
+#include <hiimjustin000.more_icons/include/MoreIcons.hpp>
 #include <IconRandomizer.hpp>
-#include <jasmine/random.hpp>
 
 using namespace geode::prelude;
 
@@ -27,8 +29,8 @@ std::map<UnlockType, std::vector<int>> unlocked = {
     { UnlockType::ShipFire, {} }
 };
 
-$on_mod(Loaded) {
-    // Initialize the Icon Randomizer API when the mod is loaded.
+$on_game(Loaded) {
+    // Initialize the Icon Randomizer API when the game is loaded.
     IconRandomizer::init();
 }
 
@@ -92,74 +94,69 @@ void Internal::addToUnlocked(UnlockType type, int id) {
     }
 }
 
-int IconRandomizer::random(int min, int max) {
-    // Get a random integer between min and max using jasmine::random.
-    return jasmine::random::getInt(min, max);
-}
-
 RandomizeType IconRandomizer::fromIconType(IconType type) {
     // Convert the icon type to a randomize type.
     switch (type) {
-        case IconType::Cube: return ICON_RANDOMIZER_API_CUBE;
-        case IconType::Ship: return ICON_RANDOMIZER_API_SHIP;
-        case IconType::Ball: return ICON_RANDOMIZER_API_BALL;
-        case IconType::Ufo: return ICON_RANDOMIZER_API_UFO;
-        case IconType::Wave: return ICON_RANDOMIZER_API_WAVE;
-        case IconType::Robot: return ICON_RANDOMIZER_API_ROBOT;
-        case IconType::Spider: return ICON_RANDOMIZER_API_SPIDER;
-        case IconType::Swing: return ICON_RANDOMIZER_API_SWING;
-        case IconType::Jetpack: return ICON_RANDOMIZER_API_JETPACK;
-        case IconType::DeathEffect: return ICON_RANDOMIZER_API_DEATH_EFFECT;
-        case IconType::Special: return ICON_RANDOMIZER_API_TRAIL;
-        case IconType::Item: return ICON_RANDOMIZER_API_ANIMATION;
-        case IconType::ShipFire: return ICON_RANDOMIZER_API_SHIP_FIRE;
-        default: return ICON_RANDOMIZER_API_CUBE;
+        case IconType::Cube: return RandomizeType::Cube;
+        case IconType::Ship: return RandomizeType::Ship;
+        case IconType::Ball: return RandomizeType::Ball;
+        case IconType::Ufo: return RandomizeType::Ufo;
+        case IconType::Wave: return RandomizeType::Wave;
+        case IconType::Robot: return RandomizeType::Robot;
+        case IconType::Spider: return RandomizeType::Spider;
+        case IconType::Swing: return RandomizeType::Swing;
+        case IconType::Jetpack: return RandomizeType::Jetpack;
+        case IconType::DeathEffect: return RandomizeType::DeathEffect;
+        case IconType::Special: return RandomizeType::Trail;
+        case IconType::Item: return RandomizeType::Animation;
+        case IconType::ShipFire: return RandomizeType::ShipFire;
+        default: return RandomizeType::Cube;
     }
 }
 
 RandomizeType IconRandomizer::fromUnlockType(UnlockType type) {
     // Convert the unlock type to a randomize type.
     switch (type) {
-        case UnlockType::Cube: return ICON_RANDOMIZER_API_CUBE;
-        case UnlockType::Col1: return ICON_RANDOMIZER_API_COLOR_1;
-        case UnlockType::Col2: return ICON_RANDOMIZER_API_COLOR_2;
-        case UnlockType::Ship: return ICON_RANDOMIZER_API_SHIP;
-        case UnlockType::Ball: return ICON_RANDOMIZER_API_BALL;
-        case UnlockType::Bird: return ICON_RANDOMIZER_API_UFO;
-        case UnlockType::Dart: return ICON_RANDOMIZER_API_WAVE;
-        case UnlockType::Robot: return ICON_RANDOMIZER_API_ROBOT;
-        case UnlockType::Spider: return ICON_RANDOMIZER_API_SPIDER;
-        case UnlockType::Streak: return ICON_RANDOMIZER_API_TRAIL;
-        case UnlockType::Death: return ICON_RANDOMIZER_API_DEATH_EFFECT;
-        case UnlockType::GJItem: return ICON_RANDOMIZER_API_ANIMATION;
-        case UnlockType::Swing: return ICON_RANDOMIZER_API_SWING;
-        case UnlockType::Jetpack: return ICON_RANDOMIZER_API_JETPACK;
-        case UnlockType::ShipFire: return ICON_RANDOMIZER_API_SHIP_FIRE;
-        default: return ICON_RANDOMIZER_API_CUBE;
+        case UnlockType::Cube: return RandomizeType::Cube;
+        case UnlockType::Col1: return RandomizeType::Color1;
+        case UnlockType::Col2: return RandomizeType::Color2;
+        case UnlockType::Ship: return RandomizeType::Ship;
+        case UnlockType::Ball: return RandomizeType::Ball;
+        case UnlockType::Bird: return RandomizeType::Ufo;
+        case UnlockType::Dart: return RandomizeType::Wave;
+        case UnlockType::Robot: return RandomizeType::Robot;
+        case UnlockType::Spider: return RandomizeType::Spider;
+        case UnlockType::Streak: return RandomizeType::Trail;
+        case UnlockType::Death: return RandomizeType::DeathEffect;
+        case UnlockType::GJItem: return RandomizeType::Animation;
+        case UnlockType::Swing: return RandomizeType::Swing;
+        case UnlockType::Jetpack: return RandomizeType::Jetpack;
+        case UnlockType::ShipFire: return RandomizeType::ShipFire;
+        default: return RandomizeType::Cube;
     }
 }
 
 IconType IconRandomizer::toIconType(RandomizeType type) {
     // Convert the randomize type to an icon type.
     switch (type) {
-        case ICON_RANDOMIZER_API_COLOR_1: return IconType::Item;
-        case ICON_RANDOMIZER_API_COLOR_2: return IconType::Item;
-        case ICON_RANDOMIZER_API_GLOW_COLOR: return IconType::Item;
-        case ICON_RANDOMIZER_API_GLOW: return IconType::Item;
-        case ICON_RANDOMIZER_API_CUBE: return IconType::Cube;
-        case ICON_RANDOMIZER_API_SHIP: return IconType::Ship;
-        case ICON_RANDOMIZER_API_BALL: return IconType::Ball;
-        case ICON_RANDOMIZER_API_UFO: return IconType::Ufo;
-        case ICON_RANDOMIZER_API_WAVE: return IconType::Wave;
-        case ICON_RANDOMIZER_API_ROBOT: return IconType::Robot;
-        case ICON_RANDOMIZER_API_SPIDER: return IconType::Spider;
-        case ICON_RANDOMIZER_API_SWING: return IconType::Swing;
-        case ICON_RANDOMIZER_API_JETPACK: return IconType::Jetpack;
-        case ICON_RANDOMIZER_API_DEATH_EFFECT: return IconType::DeathEffect;
-        case ICON_RANDOMIZER_API_TRAIL: return IconType::Special;
-        case ICON_RANDOMIZER_API_ANIMATION: return IconType::Item;
-        case ICON_RANDOMIZER_API_SHIP_FIRE: return IconType::ShipFire;
-        case ICON_RANDOMIZER_API_EXPLODE: return IconType::DeathEffect;
+        case RandomizeType::Color1: return IconType::Item;
+        case RandomizeType::Color2: return IconType::Item;
+        case RandomizeType::GlowColor: return IconType::Item;
+        case RandomizeType::Glow: return IconType::Item;
+        case RandomizeType::Cube: return IconType::Cube;
+        case RandomizeType::Ship: return IconType::Ship;
+        case RandomizeType::Ball: return IconType::Ball;
+        case RandomizeType::Ufo: return IconType::Ufo;
+        case RandomizeType::Wave: return IconType::Wave;
+        case RandomizeType::Robot: return IconType::Robot;
+        case RandomizeType::Spider: return IconType::Spider;
+        case RandomizeType::Swing: return IconType::Swing;
+        case RandomizeType::Jetpack: return IconType::Jetpack;
+        case RandomizeType::DeathEffect: return IconType::DeathEffect;
+        case RandomizeType::Trail: return IconType::Special;
+        case RandomizeType::Animation: return IconType::Item;
+        case RandomizeType::ShipFire: return IconType::ShipFire;
+        case RandomizeType::Explode: return IconType::DeathEffect;
         default: return IconType::Cube;
     }
 }
@@ -167,40 +164,40 @@ IconType IconRandomizer::toIconType(RandomizeType type) {
 UnlockType IconRandomizer::toUnlockType(RandomizeType type) {
     // Convert the randomize type to an unlock type.
     switch (type) {
-        case ICON_RANDOMIZER_API_COLOR_1: return UnlockType::Col1;
-        case ICON_RANDOMIZER_API_COLOR_2: return UnlockType::Col2;
-        case ICON_RANDOMIZER_API_GLOW_COLOR: return UnlockType::Col2;
-        case ICON_RANDOMIZER_API_GLOW: return UnlockType::Streak;
-        case ICON_RANDOMIZER_API_CUBE: return UnlockType::Cube;
-        case ICON_RANDOMIZER_API_SHIP: return UnlockType::Ship;
-        case ICON_RANDOMIZER_API_BALL: return UnlockType::Ball;
-        case ICON_RANDOMIZER_API_UFO: return UnlockType::Bird;
-        case ICON_RANDOMIZER_API_WAVE: return UnlockType::Dart;
-        case ICON_RANDOMIZER_API_ROBOT: return UnlockType::Robot;
-        case ICON_RANDOMIZER_API_SPIDER: return UnlockType::Spider;
-        case ICON_RANDOMIZER_API_SWING: return UnlockType::Swing;
-        case ICON_RANDOMIZER_API_JETPACK: return UnlockType::Jetpack;
-        case ICON_RANDOMIZER_API_DEATH_EFFECT: return UnlockType::Death;
-        case ICON_RANDOMIZER_API_TRAIL: return UnlockType::Streak;
-        case ICON_RANDOMIZER_API_SHIP_FIRE: return UnlockType::ShipFire;
-        case ICON_RANDOMIZER_API_ANIMATION: return UnlockType::GJItem;
-        case ICON_RANDOMIZER_API_EXPLODE: return UnlockType::Death;
+        case RandomizeType::Color1: return UnlockType::Col1;
+        case RandomizeType::Color2: return UnlockType::Col2;
+        case RandomizeType::GlowColor: return UnlockType::Col2;
+        case RandomizeType::Glow: return UnlockType::Streak;
+        case RandomizeType::Cube: return UnlockType::Cube;
+        case RandomizeType::Ship: return UnlockType::Ship;
+        case RandomizeType::Ball: return UnlockType::Ball;
+        case RandomizeType::Ufo: return UnlockType::Bird;
+        case RandomizeType::Wave: return UnlockType::Dart;
+        case RandomizeType::Robot: return UnlockType::Robot;
+        case RandomizeType::Spider: return UnlockType::Spider;
+        case RandomizeType::Swing: return UnlockType::Swing;
+        case RandomizeType::Jetpack: return UnlockType::Jetpack;
+        case RandomizeType::DeathEffect: return UnlockType::Death;
+        case RandomizeType::Trail: return UnlockType::Streak;
+        case RandomizeType::ShipFire: return UnlockType::ShipFire;
+        case RandomizeType::Animation: return UnlockType::GJItem;
+        case RandomizeType::Explode: return UnlockType::Death;
         default: return UnlockType::Cube;
     }
 }
 
 int IconRandomizer::randomize(RandomizeType type, bool dual) {
     // If the type is invalid, return -1.
-    if (type < 0 || type > 16) return -1;
+    if (type < RandomizeType::Color1 || type > RandomizeType::Explode) return -1;
 
     // Get the UnlockType from the randomize type, and get the vector of unlocked icons.
     auto& vec = unlocked[toUnlockType(type)];
 
     // If it is an animation, randomly enable or disable the animations, and return 0.
-    if (type == ICON_RANDOMIZER_API_ANIMATION) {
+    if (type == RandomizeType::Animation) {
         auto gameStatsManager = GameStatsManager::get();
         for (int i = 0; i < vec.size(); i++) {
-            gameStatsManager->toggleEnableItem(UnlockType::GJItem, vec[i], random(0, 1));
+            gameStatsManager->toggleEnableItem(UnlockType::GJItem, vec[i], random::generate<bool>());
         }
         return 0;
     }
@@ -214,130 +211,140 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
     auto iconType = toIconType(type);
     auto loadedIcons = more_icons::getIcons(iconType);
     auto vecSize = vec.size();
-    auto idx = random(0, vecSize + (loadedIcons ? loadedIcons->size() : 0) - 1);
+    auto idx = random::generate(0uz, vecSize + (loadedIcons ? loadedIcons->size() : 0) - 1);
     auto num = idx < vecSize ? vec[idx] : idx - vecSize;
-    auto glow = type == ICON_RANDOMIZER_API_GLOW && std::ranges::contains(unlocked[UnlockType::Streak], 2) ? (bool)random(0, 1) : false;
-    auto explode = type == ICON_RANDOMIZER_API_EXPLODE ? (bool)random(0, 1) : false;
+    auto glow = type == RandomizeType::Glow && std::ranges::contains(unlocked[UnlockType::Streak], 2) ? random::generate<bool>() : false;
+    auto explode = type == RandomizeType::Explode ? random::generate<bool>() : false;
 
     // Then, set the icon to the random number.
     // If Separate Dual Icons is enabled and the dual parameter is true, set the value in the mod.
     // Otherwise, set the value in the game manager.
     switch (type) {
-        case ICON_RANDOMIZER_API_COLOR_1:
+        case RandomizeType::Color1:
             if (useDual) separateDualIcons->setSavedValue("color1", num);
             else gameManager->setPlayerColor(num);
             return num;
-        case ICON_RANDOMIZER_API_COLOR_2:
+        case RandomizeType::Color2:
             if (useDual) separateDualIcons->setSavedValue("color2", num);
             else gameManager->setPlayerColor2(num);
             return num;
-        case ICON_RANDOMIZER_API_GLOW_COLOR:
+        case RandomizeType::GlowColor:
             if (useDual) separateDualIcons->setSavedValue("colorglow", num);
             else gameManager->setPlayerColor3(num);
             return num;
-        case ICON_RANDOMIZER_API_GLOW:
+        case RandomizeType::Glow:
             if (useDual) separateDualIcons->setSavedValue("glow", glow);
             else gameManager->setPlayerGlow(glow);
             return glow;
-        case ICON_RANDOMIZER_API_CUBE:
+        case RandomizeType::Cube:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("cube", num);
                 else gameManager->setPlayerFrame(num);
-                more_icons::setIcon("", IconType::Cube, useDual);
+                more_icons::setIcon(nullptr, IconType::Cube, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Cube, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Cube, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_SHIP:
+        case RandomizeType::Ship:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("ship", num);
                 else gameManager->setPlayerShip(num);
-                more_icons::setIcon("", IconType::Ship, useDual);
+                more_icons::setIcon(nullptr, IconType::Ship, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Ship, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Ship, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_BALL:
+        case RandomizeType::Ball:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("roll", num);
                 else gameManager->setPlayerBall(num);
-                more_icons::setIcon("", IconType::Ball, useDual);
+                more_icons::setIcon(nullptr, IconType::Ball, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Ball, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Ball, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_UFO:
+        case RandomizeType::Ufo:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("bird", num);
                 else gameManager->setPlayerBird(num);
-                more_icons::setIcon("", IconType::Ufo, useDual);
+                more_icons::setIcon(nullptr, IconType::Ufo, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Ufo, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Ufo, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_WAVE:
+        case RandomizeType::Wave:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("dart", num);
                 else gameManager->setPlayerDart(num);
-                more_icons::setIcon("", IconType::Wave, useDual);
+                more_icons::setIcon(nullptr, IconType::Wave, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Wave, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Wave, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_ROBOT:
+        case RandomizeType::Robot:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("robot", num);
                 else gameManager->setPlayerRobot(num);
-                more_icons::setIcon("", IconType::Robot, useDual);
+                more_icons::setIcon(nullptr, IconType::Robot, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Robot, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Robot, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_SPIDER:
+        case RandomizeType::Spider:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("spider", num);
                 else gameManager->setPlayerSpider(num);
-                more_icons::setIcon("", IconType::Spider, useDual);
+                more_icons::setIcon(nullptr, IconType::Spider, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Spider, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Spider, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_SWING:
+        case RandomizeType::Swing:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("swing", num);
                 else gameManager->setPlayerSwing(num);
-                more_icons::setIcon("", IconType::Swing, useDual);
+                more_icons::setIcon(nullptr, IconType::Swing, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Swing, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Swing, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_JETPACK:
+        case RandomizeType::Jetpack:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("jetpack", num);
                 else gameManager->setPlayerJetpack(num);
-                more_icons::setIcon("", IconType::Jetpack, useDual);
+                more_icons::setIcon(nullptr, IconType::Jetpack, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Jetpack, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Jetpack, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_DEATH_EFFECT:
-            if (useDual) separateDualIcons->setSavedValue("death", num);
-            else gameManager->setPlayerDeathEffect(num);
+        case RandomizeType::DeathEffect:
+            if (idx < vecSize) {
+                if (useDual) separateDualIcons->setSavedValue("death", num);
+                else gameManager->setPlayerDeathEffect(num);
+                more_icons::setIcon(nullptr, IconType::DeathEffect, useDual);
+                return num;
+            }
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::DeathEffect, useDual);
             return num;
-        case ICON_RANDOMIZER_API_TRAIL:
+        case RandomizeType::Trail:
             if (idx < vecSize) {
                 if (useDual) separateDualIcons->setSavedValue("trail", num);
                 else gameManager->setPlayerStreak(num);
-                more_icons::setIcon("", IconType::Special, useDual);
+                more_icons::setIcon(nullptr, IconType::Special, useDual);
                 return num;
             }
-            if (loadedIcons) more_icons::setIcon((*loadedIcons)[num].name, IconType::Special, useDual);
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::Special, useDual);
             return 0;
-        case ICON_RANDOMIZER_API_SHIP_FIRE:
-            if (useDual) separateDualIcons->setSavedValue("shiptrail", num);
-            else gameManager->setPlayerShipStreak(num);
+        case RandomizeType::ShipFire:
+            if (idx < vecSize) {
+                if (useDual) separateDualIcons->setSavedValue("shiptrail", num);
+                else gameManager->setPlayerShipStreak(num);
+                more_icons::setIcon(nullptr, IconType::ShipFire, useDual);
+                return num;
+            }
+            if (loadedIcons) more_icons::setIcon(loadedIcons->data() + num, IconType::ShipFire, useDual);
             return num;
-        case ICON_RANDOMIZER_API_EXPLODE:
+        case RandomizeType::Explode:
             if (useDual) separateDualIcons->setSavedValue("deathexplode", explode);
             else gameManager->setGameVariable(GameVar::PlayerExplode, explode);
             return explode;
@@ -349,29 +356,29 @@ int IconRandomizer::randomize(RandomizeType type, bool dual) {
 void IconRandomizer::randomizeAll(RandomizeAllType type, bool dual) {
     // Switch on the randomize all type and randomize all the icons of that type.
     switch (type) {
-        case ICON_RANDOMIZER_API_ALL_ICONS:
-            randomize(ICON_RANDOMIZER_API_CUBE, dual);
-            randomize(ICON_RANDOMIZER_API_SHIP, dual);
-            randomize(ICON_RANDOMIZER_API_BALL, dual);
-            randomize(ICON_RANDOMIZER_API_UFO, dual);
-            randomize(ICON_RANDOMIZER_API_WAVE, dual);
-            randomize(ICON_RANDOMIZER_API_ROBOT, dual);
-            randomize(ICON_RANDOMIZER_API_SPIDER, dual);
-            randomize(ICON_RANDOMIZER_API_SWING, dual);
-            randomize(ICON_RANDOMIZER_API_JETPACK, dual);
+        case RandomizeAllType::Icons:
+            randomize(RandomizeType::Cube, dual);
+            randomize(RandomizeType::Ship, dual);
+            randomize(RandomizeType::Ball, dual);
+            randomize(RandomizeType::Ufo, dual);
+            randomize(RandomizeType::Wave, dual);
+            randomize(RandomizeType::Robot, dual);
+            randomize(RandomizeType::Spider, dual);
+            randomize(RandomizeType::Swing, dual);
+            randomize(RandomizeType::Jetpack, dual);
             break;
-        case ICON_RANDOMIZER_API_ALL_SPECIAL:
-            randomize(ICON_RANDOMIZER_API_DEATH_EFFECT, dual);
-            randomize(ICON_RANDOMIZER_API_TRAIL, dual);
-            randomize(ICON_RANDOMIZER_API_SHIP_FIRE, dual);
-            randomize(ICON_RANDOMIZER_API_ANIMATION, dual);
-            randomize(ICON_RANDOMIZER_API_EXPLODE, dual);
+        case RandomizeAllType::Special:
+            randomize(RandomizeType::DeathEffect, dual);
+            randomize(RandomizeType::Trail, dual);
+            randomize(RandomizeType::ShipFire, dual);
+            randomize(RandomizeType::Animation, dual);
+            randomize(RandomizeType::Explode, dual);
             break;
-        case ICON_RANDOMIZER_API_ALL_COLORS:
-            randomize(ICON_RANDOMIZER_API_COLOR_1, dual);
-            randomize(ICON_RANDOMIZER_API_COLOR_2, dual);
-            randomize(ICON_RANDOMIZER_API_GLOW_COLOR, dual);
-            randomize(ICON_RANDOMIZER_API_GLOW, dual);
+        case RandomizeAllType::Colors:
+            randomize(RandomizeType::Color1, dual);
+            randomize(RandomizeType::Color2, dual);
+            randomize(RandomizeType::GlowColor, dual);
+            randomize(RandomizeType::Glow, dual);
             break;
     }
 }
@@ -379,47 +386,46 @@ void IconRandomizer::randomizeAll(RandomizeAllType type, bool dual) {
 int IconRandomizer::active(RandomizeType type, bool dual) {
     // Get the game manager, then get the Separate Dual Icons mod.
     // If the mod is enabled and the dual parameter is true, use separate dual icons.
-    auto gameManager = GameManager::get();
-    auto separateDualIcons = Loader::get()->getLoadedMod("weebify.separate_dual_icons");
-    auto useDual = separateDualIcons && dual;
+    auto separateDualIcons = dual ? Loader::get()->getLoadedMod("weebify.separate_dual_icons") : nullptr;
+    auto gameManager = separateDualIcons ? nullptr : GameManager::get();
 
     // Get the value from the Separate Dual Icons mod if it is enabled and the dual parameter is true,
     // otherwise get the value from the game manager.
     switch (type) {
-        case ICON_RANDOMIZER_API_COLOR_1:
-            return useDual ? separateDualIcons->getSavedValue("color1", 0) : gameManager->getPlayerColor();
-        case ICON_RANDOMIZER_API_COLOR_2:
-            return useDual ? separateDualIcons->getSavedValue("color2", 0) : gameManager->getPlayerColor2();
-        case ICON_RANDOMIZER_API_GLOW_COLOR:
-            return useDual ? separateDualIcons->getSavedValue("colorglow", 0) : gameManager->getPlayerGlowColor();
-        case ICON_RANDOMIZER_API_GLOW:
-            return useDual ? separateDualIcons->getSavedValue("glow", false) : gameManager->getPlayerGlow();
-        case ICON_RANDOMIZER_API_CUBE:
-            return useDual ? separateDualIcons->getSavedValue("cube", 1) : gameManager->getPlayerFrame();
-        case ICON_RANDOMIZER_API_SHIP:
-            return useDual ? separateDualIcons->getSavedValue("ship", 1) : gameManager->getPlayerShip();
-        case ICON_RANDOMIZER_API_BALL:
-            return useDual ? separateDualIcons->getSavedValue("roll", 1) : gameManager->getPlayerBall();
-        case ICON_RANDOMIZER_API_UFO:
-            return useDual ? separateDualIcons->getSavedValue("bird", 1) : gameManager->getPlayerBird();
-        case ICON_RANDOMIZER_API_WAVE:
-            return useDual ? separateDualIcons->getSavedValue("dart", 1) : gameManager->getPlayerDart();
-        case ICON_RANDOMIZER_API_ROBOT:
-            return useDual ? separateDualIcons->getSavedValue("robot", 1) : gameManager->getPlayerRobot();
-        case ICON_RANDOMIZER_API_SPIDER:
-            return useDual ? separateDualIcons->getSavedValue("spider", 1) : gameManager->getPlayerSpider();
-        case ICON_RANDOMIZER_API_SWING:
-            return useDual ? separateDualIcons->getSavedValue("swing", 1) : gameManager->getPlayerSwing();
-        case ICON_RANDOMIZER_API_JETPACK:
-            return useDual ? separateDualIcons->getSavedValue("jetpack", 1) : gameManager->getPlayerJetpack();
-        case ICON_RANDOMIZER_API_DEATH_EFFECT:
-            return useDual ? separateDualIcons->getSavedValue("death", 1) : gameManager->getPlayerDeathEffect();
-        case ICON_RANDOMIZER_API_TRAIL:
-            return useDual ? separateDualIcons->getSavedValue("trail", 1) : gameManager->getPlayerStreak();
-        case ICON_RANDOMIZER_API_SHIP_FIRE:
-            return useDual ? separateDualIcons->getSavedValue("shiptrail", 1) : gameManager->getPlayerShipFire();
-        case ICON_RANDOMIZER_API_EXPLODE:
-            return useDual ? separateDualIcons->getSavedValue("deathexplode", 0) : gameManager->getGameVariable(GameVar::PlayerExplode);
+        case RandomizeType::Color1:
+            return separateDualIcons ? separateDualIcons->getSavedValue("color1", 0) : gameManager->getPlayerColor();
+        case RandomizeType::Color2:
+            return separateDualIcons ? separateDualIcons->getSavedValue("color2", 0) : gameManager->getPlayerColor2();
+        case RandomizeType::GlowColor:
+            return separateDualIcons ? separateDualIcons->getSavedValue("colorglow", 0) : gameManager->getPlayerGlowColor();
+        case RandomizeType::Glow:
+            return separateDualIcons ? separateDualIcons->getSavedValue("glow", false) : gameManager->getPlayerGlow();
+        case RandomizeType::Cube:
+            return separateDualIcons ? separateDualIcons->getSavedValue("cube", 1) : gameManager->getPlayerFrame();
+        case RandomizeType::Ship:
+            return separateDualIcons ? separateDualIcons->getSavedValue("ship", 1) : gameManager->getPlayerShip();
+        case RandomizeType::Ball:
+            return separateDualIcons ? separateDualIcons->getSavedValue("roll", 1) : gameManager->getPlayerBall();
+        case RandomizeType::Ufo:
+            return separateDualIcons ? separateDualIcons->getSavedValue("bird", 1) : gameManager->getPlayerBird();
+        case RandomizeType::Wave:
+            return separateDualIcons ? separateDualIcons->getSavedValue("dart", 1) : gameManager->getPlayerDart();
+        case RandomizeType::Robot:
+            return separateDualIcons ? separateDualIcons->getSavedValue("robot", 1) : gameManager->getPlayerRobot();
+        case RandomizeType::Spider:
+            return separateDualIcons ? separateDualIcons->getSavedValue("spider", 1) : gameManager->getPlayerSpider();
+        case RandomizeType::Swing:
+            return separateDualIcons ? separateDualIcons->getSavedValue("swing", 1) : gameManager->getPlayerSwing();
+        case RandomizeType::Jetpack:
+            return separateDualIcons ? separateDualIcons->getSavedValue("jetpack", 1) : gameManager->getPlayerJetpack();
+        case RandomizeType::DeathEffect:
+            return separateDualIcons ? separateDualIcons->getSavedValue("death", 1) : gameManager->getPlayerDeathEffect();
+        case RandomizeType::Trail:
+            return separateDualIcons ? separateDualIcons->getSavedValue("trail", 1) : gameManager->getPlayerStreak();
+        case RandomizeType::ShipFire:
+            return separateDualIcons ? separateDualIcons->getSavedValue("shiptrail", 1) : gameManager->getPlayerShipFire();
+        case RandomizeType::Explode:
+            return separateDualIcons ? separateDualIcons->getSavedValue("deathexplode", 0) : gameManager->getGameVariable(GameVar::PlayerExplode);
         default:
             return -1;
     }